requirements.txt
wget==3.2
pandas_gbq==0.14.1
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
SQLAlchemy==1.3.23
pyspark==3.1.1
apache-airflow-providers-google==2.1.0
apache-airflow-providers-apache-spark==1.0.2
apache-airflow-providers-apache-sqoop==1.0.1
apache-airflow-providers-mongo==1.0.1
google-cloud-bigquery-storage==2.4.0
google-cloud-bigquery